+++ 3. AudioAcousticAssistant-App/Php + Sql/user.sql
@@ -1,0 +1,11 @@
+<div class="error"><h1>Error</h1><p><strong>SQL query:</strong>
+<a href="tbl_sql.php?sql_query=SET+time_zone+%3D+%22%2B00%3A00%22&amp;show_query=1&amp;db=id1289576_lake&amp;table=user&amp;token=91977939aa698e7bd337a879db646f32"><span class="nowrap"><img src="themes/dot.gif" title="Edit" alt="Edit" class="icon ic_b_edit" />&nbsp;Edit</span></a>    </p>
+<p>
+<code class="sql"><pre>
+SET time_zone = &quot;+00:00&quot;
+</pre></code>
+</p>
+<p>
+    <strong>MySQL said: </strong><a href="./url.php?url=https%3A%2F%2Fdev.mysql.com%2Fdoc%2Frefman%2F5.7%2Fen%2Ferror-messages-server.html" target="mysql_doc"><img src="themes/dot.gif" title="Documentation" alt="Documentation" class="icon ic_b_help" /></a>
+</p>
+<code>#2006 - MySQL server has gone away</code><br/></div>
