--- 3. AudioAcousticAssistant-App/Php/Login.php
@@ -1,25 +1,0 @@
-<?php
-    require("password.php");
-    $con = mysqli_connect("localhost", "id1289576_admin", "audio", "id1289576_lake");
-    
-    $username = $_POST["username"];
-    $password = $_POST["password"];
-    
-    $statement = mysqli_prepare($con, "SELECT * FROM user WHERE username = ?");
-    mysqli_stmt_bind_param($statement, "s", $username);
-    mysqli_stmt_execute($statement);
-    mysqli_stmt_store_result($statement);
-    mysqli_stmt_bind_result($statement, $colUserID, $colName, $colUsername, $colAge, $colPassword);
-    
-    $response = array();
-    $response["success"] = false;  
-    
-    while(mysqli_stmt_fetch($statement)){
-        if (password_verify($password, $colPassword)) {
-            $response["success"] = true;  
-            $response["name"] = $colName;
-            $response["age"] = $colAge;
-        }
-    }
-    echo json_encode($response);
-?>
