+++ 3. AudioAcousticAssistant-App/app/src/main/java/com/example/x15011071/audioacousticassistant_app/LoginActivity.java
@@ -7,7 +7,6 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.Toast;
 
 import com.android.volley.RequestQueue;
 import com.android.volley.Response;
@@ -65,9 +64,7 @@ public class LoginActivity extends AppCompatActivity {
                             if (success) {
 
 
-
                                 Intent intent = new Intent(LoginActivity.this, ChooseActivity.class);
-
 
 
                                 LoginActivity.this.startActivity(intent);
@@ -94,9 +91,9 @@ public class LoginActivity extends AppCompatActivity {
             }
         });
     }
-
-    @Override
+}
+/*    @Override
     public void onBackPressed() {
         Toast.makeText(getApplicationContext(),"Sorry, you can't go back here",Toast.LENGTH_LONG).show();
     }
-}
+}*/
