+++ 3. AudioAcousticAssistant-App/app/src/main/res/layout/activity_main.xml
@@ -112,18 +112,6 @@
             android:layout_width="match_parent"
             android:layout_height="75dp">
 
-            <ImageView
-
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginLeft="0dp"
-                android:layout_marginRight="0dp"
-                android:id="@+id/imageView10"
-                android:contentDescription="@string/google_signin"
-                android:padding="0dp"
-                android:scaleType="fitXY"
-                app:srcCompat="@drawable/googlesignin"
-                tools:ignore="ContentDescription" />
         </RelativeLayout>
 
     </LinearLayout>
